Discard incorrect arguments from page ranges.
When running mutool clean -da t1.pdf t4.pdf it works, but
mutool clean -da t1.pf t4 slowly tries to eat up all memory.
This is because we grow the page range list even though no
useful page range was parsed out. The fix is to print a warning
and to return NULL, triggering the caller to proceed to the next
argument.