Initialize variable even without arguments to pdf_clean_file().
Building mutool with ASAN and running mutool clean -S mydoc.pdf
causes a segmentation fault. The reason is that the len variable
is not guaranteed to be initialized, it is only set to zero if
there are arguments, i.e. argc > 0. This fix moves the
initialization outside the loop thereby fixing the crash.