Bug 707596: Handle indirect null when checking for outlines.
The attached file has a Root entry /Outlines 2 0 R, but that
object is just 2 0 obj null endobj. When parsed this becomes
an indirect PDF object which is not NULL, and then strip_outlines()
continues to try to process the expected outlines dicitionary
which fails.

The fix is to use the pdf_is_dict() convenience function to
ensure that not does /Outlines exist, but it also resolves to
the expected object type.