Squash some coverity/scanbuild issues.
1) scanbuild is concerned about qsort being called with NULL. This
can only happen when len = 0, so it's not a real issue. Add a
test to make this clear to the checker.

2) Add some 'ATLEAST' calls to execute_charstring to avoid potential
buffer/underflows.

3) Tweak dict_next to make it clear to the tools that there is no
problem (we are guaranteed that offset <= end_offset, but the tool
can't see that).

4) Fix a genuine potential problem where index_get(0) would access
out of bounds on a zero sized index.

5) Check for malformed table 'nleft' entries in charset reading code.

6) Check number of long metrics from hhea table against the number of
glyphs.

7) fz_subset_cff_for_gids is supposed to cope with being called
with a NULL buffer. Don't deref the buffer before checking it!

8) Comment out an assignment in fz_convert_edgebuffer_app that doesn't
actually get used.

9) Add a missing break in execute_charstring.

10) Add some bounds checks in subset_name_table.

11) Rejig the bounds checks in load_enc_tab4 so that the tool can
understand them.