Bug 699113: Avoid giving away borrowed reference.
ASAN reports a use after free error when running the case mentioned
in the bug report. In lpr_inherit_res_contents() an object reference
is borrowed from the parent node, but if a cycle is detected that
reference is given away by calling pdf_dict_put_drop(). This is not
allowed, instead pdf_dict_put() must be called. If no cycle is
detected a copy of the object will be created which must be passed
to pdf_dict_put_drop(). Hence the fix is to separate the two cases
as done in this commit.