Handle Annots and form XObjects that are themselves content items.
Annotations and XObjects can have a 'StructParent' field in it.
If looking this up in the ParentNums tree in the StructTreeRoot
resolves to a single entry, then the entirety of the object
should be considered part of that content element.

This addresses this for Annotations and Form XObjects, but leaves
the question of how to do it for Image XObjects unresolved.