Fix generation of invalid calRGB profiles.
When calling fz_new_icc_data_from_cal with the color data extracted
from pal8v4.bmp, the profile generated had problems, in that the
whitepoint was shifted.

This was masked in some cases by lcms2 'patching' the optimised
lookup tables it creates to ensure that the whitepoint matches.
We cannot rely on this, because not all of the routes through
lcms2 end up going through a 'patchable' pass.

Having consulted with Marti Maria Saguer, the author of lcms2,
he confirmed that the profile was being generated incorrectly,
and pointed me at the code within lcms2 that does this correctly.

I have therefore introduced new code, heavily based on the lcms2
algorithm that seems to generate the profiles correctly.

We cannot just call directly into lcms2 to have it do this work
for us, as we try to remain independent of the CMM upon which we
are built.