css: Avoid assuming that missing font file must be present in file system.
Previously if a font file specified by css was not present in the archive
MuPDF would look for that font in the file system by calling fz_read_file().
If the file does not exist fz_open_file() would fail with FZ_ERROR_SYSTEM.

Trying to read the file in the file system is speculative so it should
not be causing a fatal error. To fix this instead call fz_try_open_file()
and if the file is not present throw FZ_ERROR_FORMAT indicating that the
css rule specified a font that cannot be found.