Bug 707673: Fix text movement on sanitisation.
This is actually down to the same kind of issue as seen with
bug 707448, in that an unholy combination of negative font
height and negative font scale are combining to confuse our
positioning code.

Once again I've gone through the code attempting to simplify
it, and I think I'm happier with this version now. Certainly
it seems to get all the files from both bugs correct now.