Bug 707704: Represent JBIG2 segment length by uint_32 as per spec.
Previously the segment length was represented by a signed integer,
so when pdf_parse_jbig2_segment_header() parsed the segment length
the value read from the file could be interpreted as negative. Later
on in pdf_copy_jbig2_segments() a data pointer is incremented by
the segment length, which could lead to an out of bounds pointer and
a subsequent crash.

This fix is to represent the parsed value as an 32-bit unsigned integer
as mentioned in the specification.