Bug 707721: Avoid infinite loop in structure handling.
Our structure handling is being confused by PDF fragments that
look like this:

  <content A>
  /Foo << /MCID 0 >> BDC
  <content B>
  /Bar << /MCID 0 >> BDC
  <content C>
  EMC
  <content D>
  EMC
  <content E>

When we encounter the first BDC, we correctly begin_structure.
When we encounter the second BDC, we do not begin_structure,
because we're already in that MCID. When we hit the EMC, we
end_structure, thus unbalancing us.

This unbalancing results in us trying to pop all the way back
to the top, and a poor termination test results in us looping
with mcid_sent == NULL forever.

We fix this with various changes.

Firstly, we amend our marked_content struct with a flag saying
whether we called "begin_structure" or not. This enables us to
only "end_structure" in matching cases.

Secondly, we ensure that we only ever push dictionaries onto the
mcid_sent stack. This avoids us getting confused while popping.

Thirdly, we update the termination condition in pop_structure_to
to ensure that if we ever run off the top of the stack, we don't
loop infinitely.

Many thanks to Sebastian for doing the analysis that lead to this
fix!