Bug 707727: Fix errors when filtering type3 font.
pdf_filter_type3 makes a stack of pdf_processors to process the
character definitions from the type3 font.

We feed it the resources at the start, then process the char defs,
then read the updated resources at the end. As such, we need to
use the same stack of processors throughout - we cannot simply
recreate the processors for each character.

This means we need to 'close' the processors after each character,
and then 'reset' them again for the next one. The current code was
attempting to fudge this this and failing.

To allow us to do this neatly, we extend every pdf_processor with
a 'reopen' method that takes it back from a 'closed' state to an
operational state (broadly as if it has just been created).

This seems to solve the problem.