Bug 707727: (Prep work) Add reset method to fz_outputs.
Add the ability to 'reset' an output once it has been closed.

Not all fz_outputs will support this - those that don't will
throw an exception.

Where this is supported, it is expected to return the state
to (broadly) that which it was in after opening. It does not
'undo' any output that has already happened. Neither does it
rewind. It just returns the fz_output to a state where more
data can be written.