Bug 707733: Fix sanitize filter behaviour for empty clip paths.
If we reach an empty clip path state while sanitising, we
short-circuit various operators - there is no point in sending
content that can never be displayed.

Unfortunately, when we (I) added the code to do this, I
short-circuited the 'q' operator. I can't do that, as otherwise
'Q' operators are unbalanced.