css: Parse font shorthand property.
Note: we allow basically any form of keyword soup here, to cope with the
broken css syntax that Adobe writes in its DS strings.

The only thing we parse strictly is a font-size / line-height pair.

Abuse our relaxed font-family scanning by pointing the "font-family"
property to the start of the font list, but add a list of forbidden
keywords to avoid.