Improve recompression size checks.
We were being given a flated image (say 200K). The fz_image loads
this and holds it deflated in a buffer (say 1Meg). Then we'd
recompress this to a JPEG (say 400k), and we'd compare 400k to
1Meg and say "We should keep the new one!".

Comparing to the original /Length is better in most cases, where
this is available.

It's still not ideal because we could have an asciified stream as
input (so we'd take the input size as being larger than we should),
or we might be going out to to an asciified stream (so the output
size might be larger than we expect). BUT... broadly this is
probably good enough, because if people are interested in
recompressing images to save space, they've probably not been using
ASCII encoding anyway.