Tweak ARM architecture ifdeffery.
We use 'ARCH_ARM' to spot 32bit ARM stuff. Don't trigger this on
the presence of NEON support, because this will attempt to build
32bit ARM code when we are building for ARM64.

Conversely, don't trigger the change of weight type used for
NEON optimised deskewing code on ARCH_ARM, just trigger it for
ARCH_HAS_NEON as that's the case where it's needed.