Tweak embedded file support.
A file specifier contains the filename for a file in one
of several possible places (/F, /Unix, /DOS, /Mac). Several
of these can be set, and we pick the first one in the
arbitrary order that we use.

If a file is embedded, then it will have an /EF dictionary
entry, which will include matching versions of /F, /Unix, etc
that point to the embedded file.

It's important that if we use /F, we should look at the /F entry
from /EF and not, say, the /DOS entry.

Consider the (admittedly strange) case where we have /F, /DOS,
and /Mac entries, but only the /Mac entry is embedded, and the
/F and /DOS entries point to files on disc. (This may be an
abuse of the spec, but it is nonetheless possible).

We therefore provide a single function (get_file_stream_and_name)
to find both the embedded file and its name at the same time,
ensuring that they are always returned properly as a pair.

We adjust the code in pdf-link.c to call this, which has the
neat effect of avoiding us doing the search twice in several
places.

This also means our arbitrary choice of which order to search
for the keys in is kept in just one place.