Handle NULL format when creating document writer.
Coverity complains that if the format is NULL when passed to
fz_new_document_writer() the pointer will be dereferenced in
is_extension() potentially causing a crash. The fix is to add
a NULL-check in is_extension() causing fz_new_document_writer()
to eventually throw an exception.

This fixes Coverity issue 362814.