Add flags to fz_format_string for thousand separators for integers.
This is posix for "use thousand separator". We implement a
non-LOCALE specific version of that here. We insert a separator
every 3 digits as expected.

We support "'", "," and "_" flags for those characters as separators
respectively.

This does not match with expectations in, say, India, but it
does make output much more readable!