scripts/wrap/: fix handling of new enum/fn pdf_zugferd_profile.
If an enum and fn have same name, we need to not SWIG %ignore the name, so do
this with `pdf_zugferd_profile` like we already do with `pdf_annot_type`. [We
usually %ignore all raw C fn names because we only want SWIG to see the C++
wrapper fns.]

Also fixed generated C# out-param code when handling MuPDF enums.