Add fz_stream_filename and fz_open_file_autodelete.
fz_stream_filename, gets a filename from a stream if possible.
Might return NULL, such as when the stream does not come from a file.

fz_open_file_autodelete, opens a file that will be deleted
as when the stream is dropped. If we fail to open, it will be
deleted too!