Add 'wants_file' to fz_document_handler structure.
If this is set, then the document handler requires an actual
file, rather than just a stream. We handle this by passing
an fz_stream in, which the document_handler can retrieve the
filename of using fz_stream_filename.

In the case where a document handler 'wants_file', but the
stream can't provide a filename, the system will create a
tempfile with the stream contents in, and pass a stream that
points to that.

Care is taken to avoid, as far as possible, making this
temporary file more than once.