Only free outlines in outline iterator insert if newly created.
The code unconditionally drops the outlines, whether it is taken
from the root or it is a newly created and populated dictionary.

This is wrong because if the root already contains outlines and
the code drops them here, then the reference owned by the root
object has been discarded without it knowing about it.

Instead the code now only drops outlines if missing and new
outlines are created.