Bug 707700: Fix recolor handling of imagemasks.
It turns out that our 'BI' operator handling in the buffer
device just drops the image if it doesn't have a compressed
buffer to work from. Here we update the code to work from a
pixmap in that case.

BUT... we shouldn't even be hitting that case. If we are asked
to rewrite an imagemask, then it's pointless asking for it to
be rewritten into a different colorspace. Just exit and leave
it as it is.