Bug 707771: Fix PDF filter operator losing content.
The W and W* operators were using 'gstate->pending.ctm' as the
transform, when they should have been using the concatenation
of that and 'gstate->sent.ctm'. Accordingly, we were incorrectly
thinking that we were clipping stuff away to nothing.

This looks to be the only place we made this mistake.