Bug 707771: Fix redaction problems.
The issue here is that the 'remove pixels from images' code is
triggering removals when it should not.

Tests indicate that this is going wrong because fz_is_quad_inside_quad()
is telling us that the image is completely covered.

The cases where this is going wrong are when one of the quads is
actually a singleton point. Examination of the code shows that
this is a case we don't cope with in fz_is_point_inside_triangle.

A triangle with all 3 points identical will claim to include any
point at all. Or, I think, any triangle with points colinear.

I don't understand the maths in the old algorithm well enough to
fix it, so I swap to different algorithm that I do understand.
The old algorithm is apparently derived from:

 https://fgiesen.wordpress.com/2013/02/06/the-barycentric-conspirac/

and the new one is from:

 https://planetcalc.com/8108/