Bug 707825: Fix Rect misalignment with annotations.
When we create an annotation (say, a square or a circle etc), we
set the rectangle, set the line width, and don't set an RD.

When we come to 'update' the annotation appearance, we then draw
the border exactly aligned to the outside of the rectangle (because
there is no RD). The rect is then recalculated, and is moved
outwards, with a new RD set so that subsequent regenerations leave
the border where it is.

This means that a newly created annotation will see its Rect move
slightly when updated. The fix implemented here is to set the
RD when we set the border width.

Whenever we adjust the border width, if the width matches the
value in the RD fields, we update those too.