Bug 707798: Correct initial graphics state in pdf-device.
When creating a new PDF device, we set the initial graphics
state incorrectly. We were setting it to white in a grey
colorspace. This means that if the first thing we draw happens
to be white, we'd not send the color, and when actually
rendered, it would come out as black.