Bug 707785: Speed up layer operation.
Operations on the layer configuration (i.e. OCGs) can quickly get
slow as numbers of OCGs grow. This is because we do a linear search
through the whole list for every new one we add.

Move instead to keeping the list sorted, so we can bsearch through
it.

This changes:

 mutool -Y -o out.pdf ocp_demo.pdf

from 7 seconds down to 0.3 in debug mode.