Bug 707829: Use a big (bogus) codespace range for remapped CMaps.
Don't use the original codespace ranges, because they won't be correct
after inverting. Since we don't use the codespace ranges for these
remapped CMaps, we don't need them to be accurate, just large enough to
not trigger errors when adding the inverted ranges to the Cmap.

Note: codespace ranges are only used to decode the multi-byte strings
in content streams as arguments to the show operator TJ &co.