scripts/wrap/state.py: fix pyodide builds.
Pyodide builds have started to fail recently because `pdf_parse_ind_obj()`'s
`int64_t*` out-param is now treated as incompatible with a `long*`.

The fix here is to change our `get_name_canonical()` to preserve typedef
`int64_t` similarly to `size_t` and not convert to the underlying `long` or
`long long` type. This appears to work fine with SWIG, which sometimes has
problems with typedefs.