Bug 707824: Emit properly nested clip and layer operations in PDF.
If BMC/EMC and q/Q pairs (that apply clipping) don't nest properly, warn
and close the EMC either immediately (if badly nested one way) or at the
next Q (if badly nested the other way).

Example 1:
	q
	BMC
	Q	(close layer before this Q)
	EMC	(ignore this EMC)

Example 2:
	BMC
	q
	EMC	(ignore this EMC)
	Q	(close layer after this Q)

Also change SVG output to emit Inkscape specific layer attributes.

---

mc_depth tracks the level of BMC/EMC nesting in the content stream.

nest_mark tracks the level of layer and clip calls that we've made to the fz_device.

If the q/Q and BMC/EMC are badly nested, we ignore some EMCs and save them
until we hit the next Q. The check in do_end_layer only ends the layer if the
device and content stream are "in sync".

If we're not in sync, we ignore the EMC. This can happen in two cases: either
we saw a Q too early, or we saw an EMC too early.

In the first case we have already popped the layer (so the EMC should be
completely ignored).

In the second case the layer started before the clip and we cannot end the
layer just yet (so ignore the EMC until we see the Q and then pdf_grestore puts
the world back into order).

The loops around the pop_clip in pdf_grestore recover the syncing. The loop
before closes any layers that need to be ended before the clip is popped (layer
started with BMC inside the clip, but ends after the clip). The loop after
closes any layers that had the EMC inside the clip but were started with BMC
before the clip mask was pushed.