Determine backdrop component count once to avoid use of uninitialized components.
The fuzzed file causes read uninitialized data in lcms' Eval4Inputs().

That data comes from the backdrop array which contains uninitialized
data in fz_draw_begin_mask() since luminosity is 1. The backdrop is
converted to grayscale from the colorfv array using as many components
as indicated by the colorspace argument also passed into
fz_draw_begin_mask().

These two arguments are passed in from begin_softmask() which in
turned obtained the backdrop color from the graphics state and the
colorspace from the relevant softmask PDF object.

The backdrop color in the graphics state was stored by
pdf_run_gs_SMask() from corresponding components from the passed-in
backdrop color array using as many components as the softmask object
indicated at that point.

The backdrop color array passed to pdf_run_gs_SMask() originated in
pdf_process_extgstate() which first initializes the backdrop color
array to black using as many components as indicated by the colorspace
from the softmask's transparency group xobject. Next, each component
of the backdrop color array is set to the corresponding component
value specified by the softmask's /BC dictionary entry.

This sounds reasonable: initialize the backdrop color to black,
override the components specified by /BC and pass the backdrop color
via the graphics state stack down to the draw device function that
handles the soft masks and uses it while rendering.

But there is a flaw exposed by the fuzzed file: the number of
components in the backdrop color array is repeatedly derived from the
softmask transparency group xobject. The reason this is problematic is
that this value may change between reads in pdf_process_extgstate()
and later on in fz_draw_begin_mask() because the PDF object contents
change due to the PDF being repaired.

This is what happens when parsing the file supplied by OSS-fuzz:

In pdf_process_extgstate() the xobject colorspace is obtained from the
softmask whose transparency group is specified via indirect object
78 0 R. According to the initial xref this object should be located at
offset 192748 in the file. Unfortunately, this is not accurate, so
when MuPDF tries to resolve and cache and therefore parse the object,
a parse error is raised.

At this point, a repair is initiated by pdf_cache_object(). Once the
repair is finished the xref indicates that the 78 0 R object is now
located at offset 185643. Unfortunately due to the fuzzed nature of
the file while attempting to repair object streams in the file a
format error exception is thrown which means that there is no attempt
to parse the 78 0 R at its new location.

That format error exception is reported and the parsing of the 78 0 R
object is assumed to be null. Because 78 0 R is supposed to be a
softmask transparency group dictionary no value pairs can be looked up
in that dictionary which means that pdf_xobject_colorspace() used to
determine the colorspace of the softmask transparency group returns
NULL. In pdf_process_extgstate() this causes the number of colorspace
components to remain at the initial value of 1.

This means that the first component of the backdrop color array is
initialized to zero. Next only the first component of the 4 element
/BC array is read and replaces the first component in the backdrop
color array.

Finally, the luminosity subtype is obtained from the softmask before
the luminosity along with the backdrop color array is passed to the
callback function pdf_run_gs_SMask().

pdf_run_gs_SMask() again attempts to determine the softmask
transparency group colorspace by calling pdf_xobject_colorspace(), and
this time the 78 0 R is resolved and can be successfully parsed at the
new location indicated by the repaired xref. The problem is that the
colorspace is CMYK and therefore that four components should be used.
pdf_run_gs_SMask() dutifully copies four components from the backdrop
color array into the graphics state stack entry; the first component
from /BC along with three uninitialized component values!

Later on, while processing the page content stream when filling a path
pdf_begin_group() is called which calls begin_softmask() which passes
the half-uninitialized backdrop array to fz_begin_mask(), along with
softmask transparency group's CMYK colorspace.

Next in fz_draw_begin_mask(), because luminosity is set, it attempts
to convert 4 components from the backdrop array to a grayscale value.
During this conversion the uninitialized last three component of the
backdrop array are accessed by lcms causing valgrind and OSS-fuzz to
complain about computations using uninitialized data.

There are several possible solutions here:

 * pdf_xobject_colorspace() could be changed so that it tries to
   reparse the softmask transparency group's colorspace if the PDF was
   repaired during the first parsing of the colorspace. This would
   mean that the correct number of components is obtained from the /BC
   entry and no component values would be left uninitialized. But this
   appears to be a band-aid solution, no better than always calling
   pdf_dict_get() twice everywhere in case a value changes between the
   not-yet repaired and repaired PDF.

 * The initial backdrop array pdf_process_extgstate() could
   be initialized to all zeros. This is a straightforward solution
   that prevents the use of any uninitialized component values, but
   the number of components read out from the backdrop array
   would still vary.

 * The backdropspace array length could be determined once
   in pdf_process_extgstate() and passed through to pdf_run_gs_SMask()
   when it saves the backdrop array in the graphics state
   stack entry. Any components not set in the backdrop array
   would remain the values of the previous graphics state (or zero if
   it is at the bottommost state on the stack). This ensures that all
   components of the backdrop are initialized and the number
   of components is determined only once while parsing the PDF. The
   drawback is that the softmask transparency group is parsed again to
   determine the number of backdrop components in
   begin_softmask(), and the colorspace used for that is reparsed
   again in pdf_run_xobject(). So the component count could still
   vary.

 * The backdrop array could be dropped from the graphics
   state stack and instead reparsed in begin_softmask() since a
   pointer to the softmask PDF object is stored in the graphics state
   anyway. This guarantees that the colorspace and the background
   color array are obtained simultaneously so that the number of
   components will always match. The drawback is that the PDF object
   needs to be repeatedly reparsed.

 * The softmask transparency group colorspace could be introduced
   into the graphics state stack. In any case, it is the colorspace's
   number of components that is used to determine the number of
   components in the backdrop array, so it is as good as a
   length value. This is also obvious from that fz_convert_color()
   takes a component array and a colorspace which indicates the number
   of components. The good thing about this solution is that the
   parsing of PDF objects remains in pdf_process_extgstate() instead
   of happening repeatedly in being_softmask(). Also, the parsing of
   the colorspace occurs only once and thus the number of components
   in the backdrop array is also determined only once. The
   drawback is that yet another thing needs to hauled around in the
   graphics state and reference counted everywhere.

This commit therefore implements the last solution.

This fixes OSS-fuzz issue 70620.