Revised version of pdf_specifics commit.
Put pdf_specifics/pdf_document_from_fz_document back to how it
was before. This is because both the Java and the JS bindings
decide very early on in the lifecycle of a document whether they
should create a Document or a PDFDocument.

This is no good for documents that might take a while for the PDF
version to become available (say, because they are waiting for
a password to be presented).

Instead, we add a new fz_document_as_pdf function that will
return a (non-borrowed) reference to a pdf_document derived from
the original. If the original is a PDF, then it'll just return
the same document. If it's a document that has an underlying
PDF representation it'll return a (different) pointer to that.
Otherwise it will return NULL.