Bug 707882: Rejig fz_shade storage of functions.
Previously we would sample the functions and store the values in
a float * 256 * (FZ_MAX_COLORS+1) sized table. This meant that
shadings were ~34k each.

Here, we change to storing the values in a float * 256 * stride
table (where stride is only as big as required - typically 4 or 5).

This saves a lot of memory.