Bug 707890: Carry over structparent information when cleaning.
We were completely omitting the structure tree when copying.
This meant that information like "ActualText" was missing,
resulting in problems when doing text extraction.

Here we copy the entirety of the Structure Tree across, and
regenerate the ParentTree so that the Page StructParents still
point to the right thing.

We do NOT cut the actual Structure Tree down, so the file remains
larger than it maybe needs to be - but it is at least correct now.