Bug707859: Tweak text extraction
We were already allowing for a slight overlap in characters
when extracting, but the test file in this bug has chars
squeezed together slightly more than we were expecting.

Consider the following: (Diagram exploded vertically for
clarity - in the test file the chars are on the same line).

  +--------+
  |        |
  |        |
  +--------+
      +--------+
      |        |
      |        |
      +--------+
      |<-s-|

's' in the diagram is 'spacing' in the code.

The existing code copes with s being negative, if its absolute
size is smaller than SPACE_DIST (0.15). In this case s is around
-3.5 which absolute is comfortably less than SPACE_MAX_DIST (0.8).

Such cases were falling into the 'just consider it as a new line'
case.