Pass advance width through device interface.
In the individual text items. This is mainly useful for
PDF where the Widths array complicates matters, when combined
with substitute fonts and/or problematic embedded fonts where
multiple characters (using different PDF widths) map to the same
glyph.

This is primarily useful for the text extraction device to have more
accurate input data to work with.