Encryption dictionaries may not be saved in object streams.
Previously the Encryption dictionaries, if they exist, would be
included in object streams. While individual objects in an object
stream are not encrypted, the data of the object stream object is
encrypted. Therefore saving the Encryption dictionary inside an
object stream makes the saved PDF unreadable, since the password
hash is stored inside an encrypted object.