Bug 708035: Support encrypted PDFs with object streams.
When PDFs are saved using object streams they will also be saved
using xref streams.

Previously, when saving a PDF, any present encryption dictionary
would only be copied over from the trailer (which will not be
written to file) to the xref stream (which will be written to
file) if saving as an incremental update.

This meant that if the input document was encrypted, then the
resulting saved PDF would have all objects encrypted (including
the object stream objects), no trailer, and an xref stream object
that lacked an encryption dictionary. Such a file cannot be
opened since there is not password hash, yet streams and strings
are encrypted.

Despite the note in the message of commit
d8d60e939c571969723e92003888e8040eec31a4 there doesn't appear to
be any reason to limit copying the encryption dictionary based on
whether incremental updates are used.