Bug 708035: Handle renumbered encryption dictionaries when saving with xref streams.
Previously the PDF writing code copied the encryption dictionary
from the document trailer into the xref stream. This works well
if the encryption dictionary is entirely embedded within the xref
stream object dictionary. If the encryption dictionary is
indirectly referenced and objects are renumbered the encryption
dictionary object is likely to change object number, hence the
original document's trailer dictionary may contain a stale
indirect object reference.

In writexrefstream() we now, just like in writexref(), refer to
the encryption dictionary object kept in the writer state to
handle renumbered encryption dictionary objects. If the original
trailer embedded the encryption dictionary, we do the same in the
saved file. If the original trailer kept a reference to the
encryption dictionary object, we create a new object in the saved
file and refer to it from the xref stream by an indirect object
reference.