Plug colorspace leak when processing ext gstate softmask.
In begin_softmask() we extract the softmask and page resources, but
crucially not the colorspace, from the graphics state and store them
in softmask_save to be dropped later. In the graphics state, the
softmask and page resources are set to NULL, while the colorspace
reference remains in the graphics state.

In pdf_run_gs_SMask() there is a condition on softmask being set for
dropping of the softmask, its colorspace and the page resources set in
the graphics state.

This means that begin_softmask() may create a situation where the
softmask is NULL, but the softmask colorspace is set in the graphics
state, and then in pdf_run_gs_SMask() the colorspace is not dropped,
but instead that reference is overwritten by the new softmask
colorspace arriving as an argument.

An option would be to separate the condition in pdf_run_gs_SMask()
into one for softmask, one for the colorspace and one for the page
resources, but that would mean that the colorspace would be dropped in
pdf_run_gs_SMask() and not be valid for as long as the softmask and
the page resources references are valid.

Therefore the best fix is to always extract the softmask colorspace in
begin_softmask() and dropping the colorspace in the higher-level
functions, just like we do for the softmask itself and the page
resources.

This fixes OSS-fuzz issue 42537571.