Bug 708032: When redacting pages, create new content stream objects, do not replace them.
In the file from the bug both page 5 and page 7 refer to the same
contents stream object, 32 0 R.

So when page 5 is redacted its resources will be renamed and its
contents stream will updated conversely. But this also replaces the
contens stream for page 7, but its resources will not be renamed.
Later on when page 7 is redacted, its already updated contents stream
now refers to resources that exist by their original names in its
resource dictionary.

If the page's contents consisted of an array of streams or if the
stream object was entirely missing a new contents stream object would
be created, otherwise the contents stream object was updated in place.
By updating the contens stream object in place one page's contents
would also change another page's contents.

The fix therefore to always create a new stream object for the
new contents stream of the redacted page.