Add stext 'grid' block.
This encapsulates information from a forthcoming table-finder
module.

This block (which only appears at the start of content) suggests a
possible grid into which subsequent content might fall.

We give a list of X ordinates, a list of Y ordinates, and
'uncertainty' figures for each of those ordinates.

An uncertainty of 0 says that there is a definite vertical (for X)
or horizontal (for Y) break here. A non-zero figure of X says that
these breaks are interrupted by 'X' items of content.

These may be taken as being indicative of the presence of spanned
cells.

A strict interpretation might choose to discard all edges with
non-zero uncertainty. Or we might accept small values (relative
perhaps to the number of lines in the table, say?). The decision
is left to the consumer, as it's probably best to defer
heuristics to as late as possible in the process.

We record the 'maximum uncertainty' in the structure too - this
being the maximum number of overlapping bits of content we ever get.
The expectation is that this might be useful to compare uncertainty
values to.