Fix pdf_drop_page_imp to NULL pointers.
When we drop a PDF page, the page now lives on as a zombie in
the doc->open list until it is reaped. Previously it was freed
instantly.

Now that zombie page can be walked as part of the processing of
the open list in pdf_sync_open_pages(), so we need to be sure
that we have set the object pointer to NULL.

In fact, we set all the newly invalid pointers to NULL to avoid
us hitting this problem again in future.

Also, pdf_page contains 2 document pointers (doc and super->doc).
fz_drop_page will NULL super->doc, but we should NULL doc too.