Bug 708053/708063: Distinguish between CFF fonts with/without CIDFont operators.
The PDF from the bug report contains a Type 0 font in 6 0 R, which
has a single descendant CIDFontType0 font in 8 0 R, whose font
descriptor in 9 0 R points to a CIDFontType0C in 10 0 R indicating
that the CIDFont is represented as CFF.

The text in the PDF from the bug report consists of
"text -> [XXXXXX]" where XXXXXX are cyrillic characters. The latin
characters reside in the CID range well separated from the CID
range of the cyrillic characters. The CFF charset performs an
identity mapping for CIDs in the latin range to equivalent GIDs,
but this is not the case for characters in the cyrillic range.

Previously MuPDF assumed in its CFF font subsetting process that
all CFF fonts must mapped CIDs through the charset to determine
GIDs to each glyph.

Since the MuPDF subsetting process retains only the glyphs used in
the PDF and those characters CIDs are mapped through the CFF
charset to the GIDs where the corresponding characters are located
for latin characters those appear in the resulting file.

The cyrillic text appears to have disappeared because the CIDs are
not mapped to the same GIDs using the charset when rendering the
PDF as was assumed during the CFF font subsetting. Essentially
we end up with a PDF where the cyrillic text CIDs point to GIDs
that are not retained, and we have retained a few glyphs at GIDs
that are not used by any PDF text object.

According to the PDF specification 1.7 5.6.3 CIDFonts, subheading
"Glyph Selection in CIDFonts" page 438: only CFF fonts that use
CIDFont operators need to map CIDs though the charset to obtain
GIDs, for CFF fonts that do not use CIDFont operators no mapping
is needed since CIDs can be used directly as GIDs.

According to The Compact Font Format Specification TN 5176 1.0,
"18 CID-keyed Fonts", page 28: CFF CIDFonts can be detected by
that the Top DICT begins with a ROS operator.

The fix is therefore to detect such a ROS operator in the CFF Top
DICT and use that to signal whether CIDs should be mapped through
the charset to obtain GIDs, or whether to assume an identity
mapping from CIDs to GIDs. Since the file in the bug report
contains no such ROS operator, no mapping using the charset takes
place when subsetting the CFF fonts, just like when rendering them,
therefore causing all used characters to be retained (and no
extraneous ones kept unneccessarily).