Bug 707989: Fix breaking numbering of HTML lists.
We number lists by using a list counter as we step through
the boxes. The idea is that when we step into another layer of
list, we should store the current value, reset it to 0,
process the next layer, and then restore the saved number.

At the moment, we are only doing this store/restore when stepping into
ordered lists. We need to do this for all different list types.