ttf: Correct offset bugs in the loca table subsetting.
The offsets in the loca table must be increasing, all offsets
must be filled in, and the length of the glyph data is the
difference between an offset and the next offset.

The previous code didn't fill in the offsets of unused glyphs
between used glyphs' offsets, this caused complaints when opening
the font in freetype or fontforge.

In addition the length computed for one glyph's data was not
applied to the remapped glyph id, but to the glyph id after that
one, causing the glyph appearance to be applied to another random
glyph.