Bug 708124: Avoid PDF parser tripping over 'null' objects in objstms.
The parser encounters an object that it hasn't read from an objstm
before, loads it and finds it's a null object. It then thinks that
because it's a NULL pointer, that it didn't actually read that
object, and throws an error.

To fix this, rewrite null 'o' objects as 'f' objects.