Fix font subset usage scanner to ignore non-embedded fonts.
Use current_font < 0 to indicate that a font has not been set, or a
non-embedded (base 14 or substitute) font is currently in use.

Reset it to -1 at the start of every Tf operation, and then set
it to the correct index when we find a font that we are subsetting.