ttf: Remember to update second half of hmtx table.
A hmtx table consists of two parts:

  * an array of entries with both advance widths and left bearing values
  * an array of entries with only left bearing values

Combined, both arrays are meant to provide horizontal metrics for all
the glyphs. The number of glyphs is parsed out from the font file's
maxp table.

The length of the first array is parsed out from the font file's hhea
table. The length of the second array is computed by taking the number of
glyphs and subtracting the number of entries in the first array.

Previously, the first array length, stored in orig_num_long_hor_metrics
and was capped by the hmtx table size, while the second array length,
stored in max16, was capped both by the hmtx table size and the number of
glyphs. This meant that max16 could end up being a lower value than
orig_num_log_hor_metrics, in which case the second array entries were
never updated.

Three changes are provided in this commit:

  * Rename array lengths in the code to long_metrics and short_metrics
    respectively.
  * Use separate counter to count the number of entries in the first and
    the second arrays.
  * Make sure to cap both array lengths by the number of glyphs in the
    font.