Add pdf_clip_page.
A reuse of the existing sanitize filter to allow us to clip content
that is completely outside a given bounding box.

We may want to add the ability to add a clip path to the page in
future, but that would require changes to the filter code, so we'll
do that in a followup if required.