cff: Parse FDSelect and FDArray and each font's Private/Local subroutines.
Non-CIDFonts contain a single font. They contain a top dict and
a list of global subroutines. The top dict may additionally
contain a Private entry which points to the font's properties and
local subroutines.

CIDFonts may contain multiple fonts. They contain a top dict and
a list of global subroutines. In contrast to non-CIDFonts, the top
dict does not contain a Private entry, but instead contains a
FDSelect and a FDArray entry. FDSelect maps each glyph id to a
font in the FDArray. The FDArray in turn contains the per-font
properties and may contain a Private entry that points to each
font's properties and local subroutines.

This commit parses the FDSelect and FDArray entries and stores
their info in the subsetting state along with per-font usage.
Later on while scanning the charstrings for subroutine usage, the
correct font's local subroutines and usage are propagated while
executing the charstrings.

With this commit MuPDF can now successfully subset the fonts in
the PDF from bug 708053.