Tweak underline detection.
Currently, we are relying on the accurate bbox for a char
and checking to see if a line crosses it.

This will go wrong both for chars that are displaced from
the origin (e.g. '^') and for underlines that don't cross
chars without descenders.

As a fix for this, we calculate an 'expanded' rectangle for
each char. We extend the rectangle down to the baseline
of the glyph, and then 25% of the size of the font below
it.

This fixes lines such as "Title 1" (underlined!), which
contains no descenders, correctly being spotted as
underlined chars.