mutool: Add support to force a repair before showing objects.
This addition was triggered by the file in bug 708114 which has
been fuzzed or manually edited. Despite this the trailer and the
Encryption dictionary 6 0 R (but no other objects) can be shown,
because the xref happens to be correct for object 6 0 R and the
object itself is not badly damaged:

$ mutool show -p PASS 708114/0001.pdf xref trailer 6

While analyzing bug 708114 this behavior was surprising since when
saving the file after cleaning, the encryption dictionary 6 0 R had
disappeared! That happens during repair, because object 3 0 R has an
array that is never terminated. So object 3 0 R is considered to
extend until the end of file, encompassing also object 6 0 R, for
which consequently repair is never attempted. Hence the xref is
never extended to include 6 0 R and thus the object simply vanishes
during repair.

With this new -r flag to force repair this behavior can now easily
be seen using:

$ mutool show -p PASS 708114/0001.pdf trailer trailer.Encrypt
$ mutool show -r -p PASS 708114/0001.pdf trailer trailer.Encrypt

where the former command can successfully show object 6 0 R and
the latter, after warnings and errors printed during repair can not.

This makes it easier to debug and understand similar situations
where objects disappear or gets reinterpreted during repair.