Fix Type3 rendering of Type3Test.pdf
The PDFA published some test cases, in pdf-differences.git. This
commit addresses the behaviour of Type3WordSpacing/Type3Test.pdf.

The new interpretation of Type3 chars with stroking/filling is that
they should inherit the stroke/fill state from the graphics state,
even when 'd1' is used. This blows a hole in caching for Type3
glyphs that both stroke and fill.