Tweak Type3 font rendering in light of new spec.
The latest PDF spec has some clarifications in it on the use of
Type3 fonts with Tr values other than 0.

Update our code to accommodate this.

Also, ensure that if the glyphs of one Type 3 use another Type 3,
that we don't pollute the flags retrieved from investigating the
top one with the flags found from the bottom one.