Fix Type3 "Direct Rendering" glyphs with patterns.
When rendering a pattern, it runs in the parent graphics state.
This goes for stuff in Type3 glyphs too.

When we do our horrible hacky "render direct" stuff for awkward
Type3 glyphs, we were failing to get this correct, so some files
were being rendered wrongly.

fts_23_2303 and fts_23_2310 are good ones to test here. One defines
a Type3 that inherits the pattern from the graphics state, the
other defines a Type3 that sets the pattern inside the char.

We fix this by passing the graphics state across if required. This
isn't pretty, but seems to work.