SText: Tweak unknown character handling.
When extracting a page, if we cannot extract a valid unicode
value for a glyph, we return 0xFFFD (the unicode replacement
value).

Already, with current MuPDF, if we use the
FZ_STEXT_USE_CID_FOR_UNKNOWN_UNICODE flag, then for such
unmappable glyphs, we return the CID instead of 0xFFFD.

This commit amends this in 2 ways.

Firstly, when we use FZ_STEXT_USE_CID_FOR_UNKNOWN_UNICODE,
whenever we return the CID instead of a unicode value, we
also set the FZ_STEXT_UNICODE_IS_CID bit in the char flags, so
callers can know for which results this was done.

Secondly, we add a new FZ_STEXT_USE_GID_FOR_UNKNOWN_UNICODE,
that does the same thing, but returns the glyph number
instead of the CID (and sets FZ_STEXT_UNICODE_IS_GID).

The glyph number can be looked up to give a glyph name by
fz_get_glyph_name().