Improve garbage collection when cleaning.
If we spot streams that are identical, retry the rest of the
garbage collection process.

The example file for bug 708210 has multiple softmasks that are
all identical. The current code fails to coalesce them unless it's
run repeatedly.

Now we reach a fixed point in just one call to mutool clean.