Add CSV writer.
This dumps the textual contents of tables found in documents.
Only table content is dumped!

Each table is prefixed with a line of the form:

 Table n,r,p,x0,y0,x1,y1

where n = the number of the table in the file (starting at 0),
r = the number of rows in the table
p = page num (starting at 0)
x0,y0,x1,y1 = the bbox of the table (fitz coords)