Table hunt: Improve oversegmentation behaviour.
If we oversegment the page, then we can split a table into two
or more parts. e.g.

 AAAAA BBBBB    CCCCC DDDDD
 EEEEE FFFFF    GGGGG HHHHH
 ..... .....    ..... .....

We want to be able to consider such tables for merging at a
higher level, but we don't want to be overeager in doing this,
as we might incorrectly merge 2 unrelated tables (or push
paragraphs into tables).

This will likely want to be improved in future, but for now,
only accept such 'higher' versions if there is no uncertainty
about rows or columns.